api-gateway/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
requests>=2.31
PyJWT>=2.8
gunicorn>=21.2
psycopg2-binary>=2.9
